+++ demo/TableRowGroup.jsx
@@ -7,12 +7,14 @@
  */
 
 import React from 'react';
+import Formatter from 'uxcore-formatter';
 import Table from '../src';
 
 const commonData = {
   title: '一级标题',
   entity: '蚂蚁金服（中国）',
   institution: '招商银行丨杭州分行',
+  money: '60000',
 };
 
 const companyData = [
@@ -40,13 +42,23 @@ const mixArray = (arr1, arr2) => {
 //* 第一列为radio的demo
 class Demo extends React.Component {
   render() {
+    const rowSelection = {
+      onSelect(record, selected, selectedRows) {
+        console.log(record, selected, selectedRows);
+      },
+      onSelectAll(selected, selectedRows) {
+        console.log(selected, selectedRows);
+      },
+      // isDisabled: rowData => true,
+    };
     const tableProps = {
       jsxcolumns: [
-        { dataKey: 'company', title: '公司', width: 200 },
-        { dataKey: 'title', title: '标题', width: 200, fixed: true },
-        { dataKey: 'entity', title: '支付实体', width: 200 },
-        { dataKey: 'institution', title: '金融机构', width: 200 },
-        { dataKey: 'person', title: '申请人', width: 200 },
+        { dataKey: 'company', title: '公司', width: '20%' },
+        { dataKey: 'title', title: '标题', width: '20%', fixed: true },
+        { dataKey: 'money', title: '金额', width: '20%', type: 'money' },
+        { dataKey: 'entity', title: '支付实体', width: '50%' },
+        { dataKey: 'institution', title: '金融机构', width: '20%' },
+        { dataKey: 'person', title: '申请人', width: '20%' },
       ],
       jsxdata: {
         data: mixArray(personData, companyData).map(item => ({ ...item, ...commonData })),
@@ -54,7 +66,32 @@ class Demo extends React.Component {
       className: 'kuma-uxtable-split-line',
       rowGroupKey: 'company',
       showColumnPicker: true,
+      footer: ({ data, column, from, rowGroupData = [] }) => {
+        if (column.dataKey === 'title') {
+          return '合计';
+        }
+        if (column.dataKey === 'money') {
+          let total = 0;
+          if (from === 'rowGroup') {
+            rowGroupData.forEach((rowData) => {
+              total += parseInt(rowData.money, 10);
+            });
+          } else {
+            data.forEach((rowData) => {
+              total += parseInt(rowData.money, 10);
+            });
+          }
+
+          return Formatter.money(total.toString(), ',');
+        }
+        return null;
+      },
+      showFooter: false,
       // width: 600,
+      rowSelection,
+      size: 'small',
+      showRowGroupFooter: true,
+      height: 300,
     };
     return (
       <Table {...tableProps} />